+++ src/main/java/com/space/server/domain/ai/service/dto/response/AiResponse.java
@@ -1,4 +1,4 @@
-package com.space.server.ai.service.dto.response;
+package com.space.server.domain.ai.service.dto.response;
 
 import java.util.Map;
