+++ src/main/java/com/space/server/domain/ai/service/dto/response/gpt/tools/FileSearchTool.java
@@ -1,4 +1,4 @@
-package com.space.server.ai.service.dto.response.gpt.tools;
+package com.space.server.domain.ai.service.dto.response.gpt.tools;
 
 public record FileSearchTool(
     String type,
